[ALPS05363840] GPU: Fixing ged security issue build fail
[Detail]
Fixing ged security issue build fail

MTK-Commit-Id: 67da0f1ea6a0510ff86962f9c06c13000b33688a

CR-Id: ALPS05363840
Feature: OpenGL|ES
(cherry picked from commit acc58e4db2e35c98c93bfaad2be6cecaac3dc20d)
(cherry picked from commit d1929e557e5031f86024b052b4a3ed5fdf853e37)

+++ drivers/misc/mediatek/gpu/ged/src/ged_ge.c
@@ -374,8 +374,8 @@ int ged_bridge_ge_get(
 		psGET_IN->uint32_size * sizeof(uint32_t)) {
 		GED_PDEBUG("[%s] data (%d byte) != u32_size (%d byte)",
 			__func__,
-			(output_package_size - header_size),
-			psGET_IN->uint32_size * sizeof(uint32_t));
+			(int)(output_package_size - header_size),
+			(int)(psGET_IN->uint32_size * sizeof(uint32_t)));
 		return -EFAULT;
 	}
 
@@ -400,8 +400,8 @@ int ged_bridge_ge_set(
 		psSET_IN->uint32_size * sizeof(uint32_t)) {
 		GED_PDEBUG("[%s] data (%d byte) != u32_size (%d byte)",
 			__func__,
-			(input_package_size - header_size),
-			psSET_IN->uint32_size * sizeof(uint32_t));
+			(int)(input_package_size - header_size),
+			(int)(psSET_IN->uint32_size * sizeof(uint32_t)));
 		return -EFAULT;
 	}
